Change usize to u16

+++ src/app/mod.rs
@@ -54,7 +54,7 @@ pub struct App {
     pub rx: crossbeam::channel::Receiver<PluginAction>,
     pub whitespace_cache: Arc<Mutex<Vec<usize>>>,
     pub highlight_cache: Arc<Mutex<Vec<Vec<(Style, String)>>>>,
-    pub host_terminal_height: usize,
+    pub host_terminal_height: u16,
     pub debug_logs: Vec<String>,
 }
 
@@ -217,7 +217,7 @@ impl App {
 
     pub fn run(&mut self, term: &mut ratatui::DefaultTerminal) -> std::io::Result<()> {
         while !self.should_quit {
-            self.host_terminal_height = term.size().unwrap().height as usize;
+            self.host_terminal_height = term.size().unwrap().height;
             while let Ok(action) = self.rx.try_recv() {
                 self.dirty = true; // Mark dirty because state changed
                 match action {
